requirements.txt
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.4.1
asn1crypto==1.4.0
astroid==2.6.6
attrs==19.3.0
Automat==0.8.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.1
beautifulsoup4==4.8.2
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
coincurve==15.0.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
deluge==2.0.3
distlib==0.3.3
distro==1.4.0
Django==3.2.9
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.12.4
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.3.2
flake8==3.9.2
GeoIP==1.3.2
gevent==21.8.0
gevent-ws==2.1.0
greenlet==1.1.2
grpcio==1.16.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
importlib-metadata==4.11.0
incremental==16.10.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
mccabe==0.6.1
merkletools==1.0.3
mistune==0.8.4
more-itertools==4.2.0
msgpack==1.0.2
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nemo-emblems==5.2.0
nest-asyncio==1.5.4
netaddr==0.7.19
netifaces==0.10.4
notebook==6.4.6
numpy==1.17.4
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydocstyle==6.1.1
pyflakes==2.3.1
pygame==1.9.6
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.9.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyparted==3.11.2
pyreadstat==1.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
pysha3==1.0.2
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-libtorrent==1.1.13
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
rencode==1.0.6
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
rope==0.21.1
rsa==4.7.2
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.2.0
soupsieve==1.9.5
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
terminado==0.12.1
testpath==0.5.0
texttable==1.6.2
tinycss2==1.0.2
tldextract==2.2.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
Twisted==18.9.0
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.2.0
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.10.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
wrapt==1.12.1
xkit==0.0.0
xlrd==1.1.0
yapf==0.31.0
youtube-dl==2021.4.26
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
